structural changes and more
